:art: FORMAT (model): move project row models to root
