+++ src/models/Project/Row/ProjectRowImageModel.ts
@@ -1,4 +1,4 @@
-import AssetModel from "models/AssetModel";
+import AssetModel from "models/Asset";
 import ProjectRowModel from "./ProjectRowModel";
 
 interface ProjectRowImageModel extends ProjectRowModel {
