+++ src/models/Project/Row/ProjectRowTextModel.ts
